+++ src/eslint-plugin-coalesce/vitest.config.ts
@@ -1,0 +1,10 @@
+import { defineConfig, UserConfig } from "vitest/config";
+
+export default defineConfig({
+  test: {
+    globals: true,
+    environment: "node",
+    include: ["**/*.{test,spec}.{ts,js}"],
+    exclude: ["**/node_modules/**", "**/lib/**"],
+  },
+}) as UserConfig;